refactor: add full refactor -- new flow, audio, animations, layout etc

+++ src/webgl/entities/Item.ts
@@ -1,42 +1,65 @@
-import { Group } from "three";
+import { DoubleSide, Group, Mesh, MeshStandardMaterial, PlaneGeometry, TextureLoader } from "three";
 import gsap from "gsap";
 
 import { AudioEffects } from "@/contexts/audioContext";
 
 import { Scene } from "@/webgl/globals/Scene";
 
-import { AUDIO_PLAY_EFFECT, GL_ACTIVATE_FOCUS } from "@/webgl/config/topics";
+import { AUDIO_PLAY_EFFECT } from "@/webgl/config/topics";
 import { ItemData } from "@/webgl/config/types";
 
 export class Item {
   scene = Scene.getInstance();
 
   itemData: ItemData;
-  model: Group;
+  mesh?: Mesh;
 
-  constructor(itemData: ItemData, model: Group) {
+  constructor(itemData: ItemData) {
     this.itemData = itemData;
-    this.model = model;
 
     this.init();
   }
 
-  init() {
-    const position = this.itemData.position;
+  async init() {
+    const textureLoader = new TextureLoader();
 
-    this.model.position.set(position.x, position.y, position.z);
-    this.scene.add(this.model);
+    const texture = await textureLoader.load(
+      this.itemData.src
+    );
+
+    const geometry = new PlaneGeometry(1.44, 0.9, 1);
+    const material = new MeshStandardMaterial({
+      map: texture,
+      side: DoubleSide,
+    });
+    this.mesh = new Mesh(geometry, material);
+    this.mesh.scale.set(this.itemData.scalar, this.itemData.scalar, this.itemData.scalar);
+
+    this.mesh.position.set(
+      this.itemData.position.x,
+      this.itemData.position.y,
+      this.itemData.position.z
+    );
+    this.mesh.rotation.set(
+      this.itemData.rotation.x,
+      this.itemData.rotation.y,
+      this.itemData.rotation.z
+    );
+
+    this.scene.add(this.mesh);
   }
 
   move() {
+    if (!this.mesh) return;
+
     gsap.fromTo(
-      this.model.position,
+      this.mesh.position,
       {
-        z: 1.7,
+        z: this.itemData.position.z,
       },
       {
         duration: 2,
-        z: 2,
+        z: 2.066,
         onComplete: () => {
           this.drop();
         },
@@ -45,21 +68,19 @@ export class Item {
   }
 
   drop() {
+    if (!this.mesh) return;
     let playedEffect = false;
 
     const timeline = gsap.timeline();
     timeline.fromTo(
-      this.model.position,
+      this.mesh.position,
       {
-        y: this.model.position.y,
+        y: this.mesh.position.y,
       },
       {
         duration: 0.5,
         ease: "power4.in",
         y: -2,
-        onStart: () => {
-          PubSub.publish(GL_ACTIVATE_FOCUS);
-        },
         onUpdate: () => {
           if (timeline.progress() > 0.8 && !playedEffect) {
             playedEffect = true;
@@ -68,17 +89,21 @@ export class Item {
           }
         },
         onComplete: () => {
-          this.focus();
+          if (typeof window !== 'undefined') {
+            setTimeout(() => {
+              window.open(this.itemData.slug, '_blank');
+            }, 333);
+          }
         },
       }
     );
 
     gsap.fromTo(
-      this.model.rotation,
+      this.mesh.rotation,
       {
-        x: this.model.rotation.x,
-        y: this.model.rotation.y,
-        z: this.model.rotation.z,
+        x: this.itemData.rotation.x,
+        y: this.itemData.rotation.y,
+        z: this.itemData.rotation.z,
       },
       {
         duration: 0.5,
@@ -88,23 +113,5 @@ export class Item {
         z: 0.5,
       }
     );
-  }
-
-  focus() {
-    gsap.to(this.model.rotation, {
-      duration: 3,
-      ease: "power4.inOut",
-      x: 0,
-      y: 0,
-      z: 0,
-    });
-
-    gsap.to(this.model.position, {
-      x: 0,
-      y: 0,
-      z: 9,
-      duration: 3,
-      ease: "power4.inOut",
-    });
   }
 }
